fix: 파일명 수정,type분리,disabled, error 추가 [#10]
