+++ src/components/common-ui/Button/Button.tsx
@@ -1,26 +1,16 @@
-import React from 'react'
-import { StyledButton } from './Button.styled'
+import { StyledButton } from './Button.styles'
+import { ButtonProps } from '@/types/button'
 
 /*
 사용예시 - PR 예시 화면 참고해주세요
       <Button
         color="pink"
         size="small"
-        //fontSize="14px" (fontSize ,padding 커스텀 가능)
+        //fontSize="14px" (fontSize ,padding,disabled 커스텀 가능)
         >
         수정완료
       </Button>
 */
-type ButtonColor = 'pink' | 'gray' | 'transparent'
-type ButtonSize = 'small' | 'medium' | 'large'
-
-export interface ButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  color?: ButtonColor
-  size?: ButtonSize
-  fontSize?: string
-  padding?: string
-}
 
 const Button = ({
   color = 'pink',
