+++ src/components/common-ui/Input/Input.styles.ts
@@ -1,33 +1,35 @@
 import styled from 'styled-components'
 
-export const InputWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin-bottom: var(--space-medium);
-`
-
-export const StyledInput = styled.input<{ width?: string; fontSize?: string }>`
+export const StyledInput = styled.input<{
+  width?: string
+  fontSize?: string
+  error?: boolean
+}>`
   padding: 10px;
   font-size: ${({ fontSize }) => fontSize || 'var(--font-medium)'};
-  border: 1px solid var(--color-gray);
+  border: 1px solid
+    ${({ error }) => (error ? 'var(--color-red)' : 'var(--color-gray)')};
   background-color: transparent;
   color: var(--color-white);
   border-radius: var(--border-radius-medium);
   width: ${({ width }) => width || '300px'};
   &:focus {
     outline: none;
-    border-color: var(--color-white);
+    border-color: ${({ error }) =>
+      error ? 'var(--color-red)' : 'var(--color-white)'};
   }
 `
 
 export const StyledTextarea = styled.textarea<{
   width?: string
   fontSize?: string
+  error?: boolean
 }>`
   font-family: inherit;
   padding: 10px;
   font-size: ${({ fontSize }) => fontSize || 'var(--font-medium)'};
-  border: 1px solid var(--color-gray);
+  border: 1px solid
+    ${({ error }) => (error ? 'var(--color-red)' : 'var(--color-gray)')};
   background-color: transparent;
   color: var(--color-white);
   border-radius: var(--border-radius-medium);
@@ -35,6 +37,7 @@ export const StyledTextarea = styled.textarea<{
   width: ${({ width }) => width || '800px'};
   &:focus {
     outline: none;
-    border-color: var(--color-white);
+    border-color: ${({ error }) =>
+      error ? 'var(--color-red)' : 'var(--color-white)'};
   }
 `
