+++ src/types/button.ts
@@ -1,0 +1,10 @@
+type ButtonColor = 'pink' | 'gray' | 'transparent'
+type ButtonSize = 'small' | 'medium' | 'large'
+
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  color?: ButtonColor
+  size?: ButtonSize
+  fontSize?: string
+  padding?: string
+}
