+++ src/types/input.ts
@@ -1,0 +1,12 @@
+export interface InputProps {
+  label?: string
+  placeholder?: string
+  type?: 'text' | 'textarea'
+  value: string
+  onChange: (
+    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
+  ) => void
+  width?: string
+  fontSize?: string
+  error?: boolean
+}
