+++ src/types/profile.ts
@@ -1,0 +1,8 @@
+export type ProfileButtonSize = 'small' | 'medium' | 'large' | string
+
+export interface ProfileButtonProps {
+  imageUrl?: string
+  altText?: string
+  onClick: () => void
+  size?: ProfileButtonSize
+}
